refactored MachinerySkeletonLoader and added MachinerySkeletonLoader.css
