+++ translate/src/modules/loaders/components/SkeletonLoader.css
@@ -23,42 +23,6 @@
   display: inline-block;
 }
 
-/* Machinery skeleton */
-
-.machinery-skeleton {
-  list-style: none;
-  padding: 0;
-  margin: 0;
-}
-
-.machinery-skeleton .machinery-item-skeleton {
-  border-bottom: 1px solid var(--main-border-1);
-  padding: 10px;
-}
-
-.machinery-skeleton .machinery-item-skeleton p {
-  background-color: var(--translation-secondary-color);
-  margin-bottom: 8px;
-}
-
-.machinery-skeleton .machinery-item-skeleton p.suggestion {
-  width: 75%;
-  margin-bottom: 0;
-}
-
-.machinery-skeleton .machinery-item-skeleton header {
-  display: block;
-  padding: 1px 0 5px;
-  text-align: right;
-}
-
-.machinery-skeleton .machinery-item-skeleton header .skeleton-bar {
-  background-color: var(--translation-secondary-color);
-  display: inline-block;
-  height: 0.6875rem;
-  width: 40%;
-}
-
 @keyframes fading {
   0% {
     opacity: 0.1;
